% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{calculate_boxplot_stats}
\alias{calculate_boxplot_stats}
\title{Calculate Boxplot Summary Statistics for a Numeric Vector}
\usage{
calculate_boxplot_stats(
  numeric_vector,
  id = NULL,
  return_dataframe = TRUE,
  outliers_as_strings = FALSE,
  delim = "|"
)
}
\arguments{
\item{numeric_vector}{A numeric vector for which to calculate boxplot summary statistics.}

\item{id}{An optional string to add as an ID column to the dataframe}

\item{return_dataframe}{Logical, if TRUE, returns a dataframe; otherwise returns a list.}

\item{outliers_as_strings}{Logical. Return outliers as a character vector of '|' delimited strings rather than as a list-column (ignored if \code{return_dataframe = FALSE}).
Can customise the delimiter using the \code{delim} argument}

\item{delim}{the delimiter separating different outliers. Used when outlier column is a character vector where each element describes a set of outliers delimited by some character, usually '|'.}
}
\value{
A list or dataframe containing the following elements:
\describe{
\item{min}{The minimum value of the vector.}
\item{max}{The maximum value of the vector.}
\item{median}{The median value of the vector.}
\item{q1}{The first quartile (25th percentile) of the vector.}
\item{q3}{The third quartile (75th percentile) of the vector.}
\item{iqr}{The interquartile range, calculated as \code{q3 - q1}.}
\item{outlier_low_threshold}{The lower threshold for outliers, calculated as \code{q1 - 1.5 * iqr}.}
\item{outlier_high_threshold}{The upper threshold for outliers, calculated as \code{q3 + 1.5 * iqr}.}
\item{outliers}{A list-column of outlier values that are either below the lower threshold or above the upper threshold. Can also be a '|' delimited character vector of outliers}
\item{id}{An optional ID if provided.}
}
}
\description{
This function calculates the summary statistics typically used in a boxplot,
including the minimum, maximum, median, first quartile (Q1), third quartile (Q3),
interquartile range (IQR), and outliers.
}
\examples{
vec <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100)
calculate_boxplot_stats(vec)

# More complex version with actual outliers
calculate_boxplot_stats(c(rep(1, times =3), 1:10, 22, 23))

}
