% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{calculate_boxplot_stats_for_multiple_groups}
\alias{calculate_boxplot_stats_for_multiple_groups}
\title{Calculate Boxplot Summary Statistics for Multiple Groups}
\usage{
calculate_boxplot_stats_for_multiple_groups(values, ids)
}
\arguments{
\item{values}{A numeric vector for which to calculate boxplot summary statistics.}

\item{ids}{A vector of IDs corresponding to the groups in 'values'.}
}
\value{
A dataframe containing boxplot summary statistics for each group.
}
\description{
This function calculates boxplot summary statistics for a numeric vector,
grouped by a corresponding vector of IDs.
}
\examples{
values <- c(1, 1, 1, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 22, 23, 8, 8, 8, 12,
    13, 14, 15, 30, 31)
ids <- c("b1", "b1", "b1", "b1", "b1", "b1", "b1", "b1", "b1", "b1",
 "b1", "b1", "b1", "b1", "b1", "b2", "b2", "b2", "b2", "b2", "b2",
 "b2", "b2", "b2")

calculate_boxplot_stats_for_multiple_groups(values, ids)
}
