% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{delim_to_list_column}
\alias{delim_to_list_column}
\title{Convert Delimited String to List Column}
\usage{
delim_to_list_column(char_vector, delim = "|")
}
\arguments{
\item{char_vector}{A character vector containing delimited numeric strings to convert.}

\item{delim}{The character used as a delimiter to separate the values in the string (defaults to '|').}
}
\value{
A list column where each element is a numeric vector extracted from the delimited string.
}
\description{
This function converts a character vector with delimited strings into a list column of numeric vectors.
}
\details{
\code{delim_to_list_column(c("1|2|3", "4|5"), delim = "|")}
}
