% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{list_column_to_delim}
\alias{list_column_to_delim}
\title{Convert List Column to Delimited String}
\usage{
list_column_to_delim(list_column, delim = "|")
}
\arguments{
\item{list_column}{A list column to convert.}

\item{delim}{the character used as a delimiter to separate the outlier values (defaults to '|')}
}
\value{
A character vector with delimited strings.
}
\description{
This function converts list columns in a dataframe to delimited strings.
}
