% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{make_interactive}
\alias{make_interactive}
\title{Make boxplots interactive}
\usage{
make_interactive(ggobj, ...)
}
\arguments{
\item{ggobj}{ggplot object to print. Argument \code{code} will
be ignored if this argument is supplied.}

\item{...}{
  Arguments passed on to \code{\link[ggiraph:girafe]{ggiraph::girafe}}
  \describe{
    \item{\code{code}}{Plotting code to execute}
    \item{\code{pointsize}}{the default pointsize of plotted text in pixels, default to 12.}
    \item{\code{width_svg,height_svg}}{The width and height of the graphics region in inches.
The default values are 6 and 5 inches. This will define the aspect ratio of the
graphic as it will be used to define viewbox attribute of the SVG result.

If you use \code{girafe()} in an 'R Markdown' document, we
recommend not using these arguments so that the knitr
options \code{fig.width} and \code{fig.height} are used instead.}
    \item{\code{options}}{a list of options for girafe rendering, see
\code{\link[ggiraph:opts_tooltip]{opts_tooltip()}}, \code{\link[ggiraph:opts_hover]{opts_hover()}}, \code{\link[ggiraph:opts_selection]{opts_selection()}}, ...}
    \item{\code{dependencies}}{Additional widget HTML dependencies, see \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}.}
  }}
}
\description{
Make boxplots interactive.
}
