% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{plot_boxplot_stats}
\alias{plot_boxplot_stats}
\title{Plot Boxplot Statistics}
\usage{
plot_boxplot_stats(stats, xlab = "ID", ylab = "Value")
}
\arguments{
\item{stats}{A dataframe containing boxplot statistics.}

\item{xlab, ylab}{axis title}
}
\description{
This function creates a boxplot using ggplot2 based on the boxplot statistics dataframe.
}
