% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{unnest}
\alias{unnest}
\title{Simple Unnest Function}
\usage{
unnest(ids, nested_list)
}
\arguments{
\item{ids}{A vector of IDs corresponding to each nested list.}

\item{nested_list}{A list of vectors, where each element of the list corresponds to the values for each ID.}
}
\value{
A data frame with two columns: 'id' and 'values', where 'values' are the unnested elements of the original list.
}
\description{
This function unnests a list column in a data frame, creating a long format data frame.
}
\examples{
ids <- c(1, 2, 3)
nested_list <- list(
  c(1, 2, 3),
  c(4, 5),
  c(6, 7, 8, 9)
)
unnest(ids, nested_list)

}
