% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{write_boxplot_stats_tsv}
\alias{write_boxplot_stats_tsv}
\title{Write Boxplot Statistics to TSV File}
\usage{
write_boxplot_stats_tsv(df, file)
}
\arguments{
\item{df}{A dataframe containing boxplot statistics.}

\item{file}{The file path where the TSV file will be written.}
}
\description{
This function writes the boxplot statistics dataframe to a TSV file.
}
