% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotstats.R
\name{calculate_boxplot_stats_for_multiple_groups}
\alias{calculate_boxplot_stats_for_multiple_groups}
\title{Calculate Boxplot Summary Statistics for Multiple Groups}
\usage{
calculate_boxplot_stats_for_multiple_groups(
  values,
  ids,
  outliers_as_strings = FALSE,
  delim = "|"
)
}
\arguments{
\item{values}{A numeric vector for which to calculate boxplot summary statistics.}

\item{ids}{A vector of IDs corresponding to the groups in 'values'.}

\item{outliers_as_strings}{Logical. Return outliers as a character vector of '|' delimited strings rather than as a list-column (ignored if \code{return_dataframe = FALSE}).
Can customise the delimiter using the \code{delim} argument}

\item{delim}{the character used as a delimiter to separate the outlier values (defaults to '|')}
}
\value{
A dataframe containing boxplot summary statistics for each group.
}
\description{
This function calculates boxplot summary statistics for a numeric vector,
grouped by a corresponding vector of IDs.
}
\examples{
values <- c(1, 1, 1, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 22, 23, 8, 8, 8, 12,
    13, 14, 15, 30, 31)
ids <- c("b1", "b1", "b1", "b1", "b1", "b1", "b1", "b1", "b1", "b1",
 "b1", "b1", "b1", "b1", "b1", "b2", "b2", "b2", "b2", "b2", "b2",
 "b2", "b2", "b2")

calculate_boxplot_stats_for_multiple_groups(values, ids)
}
